Garfield-master/ChatBotAPI/requirements.txt
flask
gunicorn
aiml==0.9.2
beautifulsoup4==4.9.1
lxml
googlesearch-python
certifi==2020.6.20
chardet==3.0.4 
requests==2.24.0
soupsieve==2.0.1
urllib3==1.25.9

